requirements.txt
streamlit
pandas
tensorflow
google-cloud-aiplatform
protobuf
matplotlib